Implement the same_page encoding
When a mini-transaction is updating multiple fields in a page,
it can avoid repeating the tablespace identifier and page number.
The byte offset of the record will be relative to where the
previous record on that page ended.

mtr_t::m_last, mtr_t::m_last_offset: End of the latest m_log record.

page_recv_t::last_offset: Reflects mtr_t::m_last_offset.

Valid values for last_offset during recovery should be 0 or above 8.
(The first 8 bytes of a page are the checksum and the page number,
and neither are ever updated directly by log records.)
We assign the special value 1 to indicate that the same_page form
is not allowed for the subsequent record.